+++ CppMinecraft/src/blocks/block_type.h
@@ -18,9 +18,9 @@ namespace Blocks
 		bool is_transparent;
 		bool is_cube;
 		const char* name;
-		unsigned int id;
-		BlockType(const char* name,	unsigned int block_id, Models::Model* model, 
-			const std::vector<unsigned int>& texture_layout,
+		uint32_t id;
+		BlockType(const char* name, size_t block_id, Models::Model* model,
+			const std::vector<size_t>& texture_layout,
 			bool is_transparent, bool is_cube);
 		BlockType(const BlockType& other);
 		~BlockType() noexcept;
@@ -28,7 +28,7 @@ namespace Blocks
 		{
 			return quads;
 		}
-		unsigned int get_quad_number() const
+		size_t get_quad_number() const
 		{
 			return model->get_quad_number();
 		}
